+++ CHANGELOG.md
@@ -56,6 +56,7 @@ Compatibility:
 * Update `Kernel#instance_variables` to return insertion order (@bjfish).
 * Fixed `rb_path2class()` to not error for a module (#2511, @eregon).
 * Update `Kernel#print` to print `$_` when no arguments are given (#2531, @bjfish).
+* Add category kwarg to Kernel.warn and Warning.warn (#2533, @Strech).
 
 Performance: